+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+vUNRELEASED
+---------------
+* :+1: Added the possibility to import `Parts` from an `Excel` file (#931)
+
 v4.3.0 (7NOV22)
 ---------------
 * :+1: Added the possibility to retrieve the `form_collection` from an `Activity`, if applicable (#1172)